[METROL-480] Fix for Netflix crash on trailer end.
There is a delay in Netflix Application calling
playerClose API. There is a thread which tries to
flush presented fragments from mMediaTrackFragments
and returns the PTS for the first element from the
list. Whenever there is a delay, it would delete
all the elements and tries to fetch the PTS from
the non existing element which results in the crash.

As a fix now we are checking if the list is empty
before trying to get the PTS from the front element.