package/binutils: add support for 2.45
GCC versions 11 and 12 will emit DWARF5 debug info by default
when enabled. Our binutils does not account for that, so the
newer compilers should be coupled with a newer binutils version.

2.45 was chosen as it is the first version that does not have
DWARF errors when using objdump with WPE.

A pixman patch is included to fix build with newer binutils
versions that have a stricter assembler on ARM.